Make fabrish for fabric intermediaries
